placement: Move schematic placement code to schematic folder
