+++ docs/ref/constraints.rst
@@ -18,7 +18,7 @@ In ORD viewgens, the ``!`` statement adds a constraint to the view's own
 solver, which is solved automatically at the end of the viewgen body; no
 manual Solver is needed there. This is available in layout and schematic
 views. For placing schematic elements relative to each other without
-explicit coordinates, see :mod:`ordec.core.placement`.
+explicit coordinates, see :mod:`ordec.schematic.placement`.
 
 **Example:**
