+++ docs/ref/ord.rst
@@ -58,11 +58,18 @@ Python to supply the necessary constructs during execution.
 Node Statements
 ^^^^^^^^^^^^^^^
 
-A **node statement** is the ``A B`` construct that creates and names an element in the current context. There are three types of node statements:
+A **node statement** is the ``A B`` construct that creates and names an element in the current context. There are four types of node statements:
 
 1. **Node class statements** — the type is a Node subclass, e.g., ``LayoutRect x``
 2. **Node instance statements** — the type is a Cell class or instance, e.g., ``Nmos x``
 3. **Node keyword statements** — the type is a built-in keyword, e.g., ``input x``, ``output y``, ``port z``
+4. **Placement group statements** — the type is a :class:`~ordec.schematic.placement.PlacementGroup` subclass, e.g., ``Col(gap=4) stack``
+
+Placement group statements differ from the first three types: the group is
+recorded on the view context instead of being inserted into the subgraph, and
+its body is naming-transparent — elements declared inside are added to the
+view root as usual and only additionally registered as children of the group
+(see :mod:`ordec.schematic.placement`).
 
 A node statement may have an optional body (indented block after ``:``) for setting attributes:
 
@@ -268,6 +275,26 @@ actual connection node (``SchemInstanceConn`` or
 
 Because ``--`` is plain Python arithmetic, it coexists with regular numeric
 expressions: ``2 -- 2`` evaluates to ``4`` as expected.
+
+
+Constraint Statements ``!``
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+A statement starting with ``!`` declares a linear constraint instead of
+executing a normal statement. The constraint is added to the view's own
+:class:`~ordec.core.constraints.Solver`, which is solved automatically at the
+end of the viewgen body; this is available in layout and schematic views (see
+:mod:`ordec.core.constraints`):
+
+.. code-block::
+
+    Col(gap=4) stack:
+        Pmos pu
+        Nmos pd
+    ! stack.southwest == (3, 1)
+
+Internally, ``! expr`` compiles to ``context.constrain(expr)``, which forwards
+the constraint to the active view context.
 
 
 The following summary shows the most important functions and classes of ORD.
